replace fixed4 with float4 in shaders

+++ Visualisation/Shader/Base/RaycastCylinder.cginc
@@ -76,8 +76,8 @@ struct v2f
     float4 vertex : SV_POSITION;
     float4 q : TEXCOORD0;
     float4 d : TEXCOORD1;
-    fixed4 color1 : TEXCOORD2;
-    fixed4 color2 : TEXCOORD3;
+    float4 color1 : TEXCOORD2;
+    float4 color2 : TEXCOORD3;
     float3 a : TEXCOORD4;
 };
 
@@ -141,7 +141,7 @@ v2f vert (appdata i)
 
 
 struct fout {
-    fixed4 color : SV_Target;
+    float4 color : SV_Target;
     float depth : SV_Depth;
 };
